Relocate core primitives into their own files
Additionally use the new Scala 3 import wildcard syntax.
